+++ src/components/button.js
@@ -1,0 +1,7 @@
+/**
+ * Функция возращает html разметку кнопки загрузить еще.
+ * @return {string}
+ */
+const getButtonLoadMore = () => `<button class="load-more" type="button">load more</button>`;
+
+export {getButtonLoadMore};
